Crate noir_greybox_fuzzer

Modules

corpus 🔒

coverage 🔒

This file implements the mechanisms for coverage - detection of changes is the execution of the target program It assists in exploration of the program through testcase mutation by telling the fuzzer whether a new testcase represents previously unexplored functionality. This in turn allows the fuzzer to add them to the corpus as footholds for further exploration

dictionary 🔒

This module defines how to build a dictionary of values which are likely to be correspond to significant inputs during fuzzing by inspecting the Program being fuzzed.

mutation 🔒

types 🔒

Structs

AcirAndBrilligPrograms

FastParallelFuzzResult 🔒

Contains information from parallel execution of testcases for quick single-threaded processing If no new coverage is detected, the fuzzer can simply quickly update the timing metrics without parsing the outcome

FuzzTask 🔒

A structure with the values for a single mutation-fuzz iteration in the fuzzer

FuzzedExecutor

An executor for Noir programs which which provides fuzzing support

FuzzedExecutorExecutionConfiguration

FuzzedExecutorFolderConfiguration

Metrics 🔒

The metrics of the fuzzing process being output to the user

Enums

FuzzTestResult

The outcome of a fuzz test

FuzzedExecutorFailureConfiguration

Constants

FOREIGN_CALL_FAILURE_SUBSTRING 🔒

MINIMUM_PULSE_INTERVAL_MILLIS 🔒

Minimum pulse interval in milliseconds for printing metrics

SINGLE_FUZZING_ROUND_TARGET_TIME 🔒

We aim the number of testcases per round so one round takes these many microseconds

Functions

build_dictionary_from_program

Constructs a HashSet<F> of values pulled from a Program<F> which are likely to be correspond to significant inputs during fuzzing.

display_metrics 🔒

display_starting_info 🔒

Type Aliases

ErrorAndCoverage

Information collected from testcase execution on failure in brillig

ErrorAndWitness

Information collected from testcase execution on failure in ACIR

SimpleXorShiftRNGSeed 🔒

A seed for the XorShift RNG for use during mutation

WitnessAndCoverage

Information collected from testcase execution on success