noir_greybox_fuzzer

Module coverage

This file implements the mechanisms for coverage - detection of changes is the execution of the target program It assists in exploration of the program through testcase mutation by telling the fuzzer whether a new testcase represents previously unexplored functionality. This in turn allows the fuzzer to add them to the corpus as footholds for further exploration

There are several mechanisms for coverage being used:

Standard branch coverage taken from brillig, the same as with standard, non-zk programs (detects which branch has been taken in an if)

Conditional move coverage from brillig

Novel boolean witness coverage. If ACIR execution was successful, we scan the witness for potential boolean values and detect interesting testcases, if we discover a boolean witness with a state that hasn’t been previously encountered.

Structs

AccumulatedCmpCoverage

Metrics of the closeness of a particular comparison

AccumulatedFuzzerCoverage

Total coverage presented by all testcases in the corpus

AccumulatedSingleBranchCoverage

Metrics of a particular branch

AcirBoolState

Represents a single encountered state of a boolean witness in the Acir program

BranchCoverageRange

Structure containing information that at a particular index in the brillig program there is a branch

CmpCoverageRange

Structure containing information that at a particular index in the brillig program there is a comparison between elements of the following bit size

NonBoolWitnessList

Mechanism for excluding non-boolean witnesses from the ACIR witness map This mechanism was introduced after we started to use partial witness maps from failed executions

SingleTestCaseCoverage

Information about the coverage of a single testcase execution

Enums

BrilligCoverageItemRange

Structure containing information about positions of coverage-related opcodes in the brillig program

Constants

EQ_STATE_COUNT 🔒

FUZZING_COMPARISON_FALSE_STATE 🔒

A state that represents a false comparison as part of a feature

FUZZING_COMPARISON_LOG_RANGE_NUMBER_OF_STATES 🔒

The number of states in the range of the states that represent logarithm of the difference between the comparison arguments as part of a feature

FUZZING_COMPARISON_LOG_RANGE_START_STATE 🔒

The start of the range of the states that represent logarithm of the difference between the comparison arguments as part of a feature

FUZZING_COMPARISON_TRUE_STATE 🔒

A state that represents a true comparison as part of a feature

IF_BRANCH_COUNT 🔒

Functions

analyze_brillig_program_before_fuzzing

Analyze the brillig program to detect:

Type Aliases

BrilligCoverageRanges

Feature

A tuple of the current opcode position and the next opcode position or state

FeatureToIndexMap

A map from a particular branch or comparison to its unique index in the raw vector used inside brillig vm

NextOpcodePositionOrState

The position of the next opcode that will be executed in the bytecode or an id of a specific state produced by the opcode

OpcodePosition

The position of an opcode that is currently being executed in the bytecode

RawBrilligCoverage

Raw brillig coverage is just a buffer of unsigned ints that contain counters

UniqueFeatureIndex

The index of a unique feature in the fuzzing trace

UnusedTestcaseIdSet 🔒