noir_greybox_fuzzer
Module dictionary

This module defines how to build a dictionary of values which are likely to be correspond to significant inputs during fuzzing by inspecting the [Program] being fuzzed.

This dictionary can be fed into the fuzzer’s [strategy][proptest::strategy::Strategy] in order to bias it towards generating these values to ensure they get proper coverage.

Functions

build_dictionary_from_circuit 🔒

Collect Field values used in the opcodes of an ACIR circuit.

build_dictionary_from_program

Constructs a HashSet<F> of values pulled from a [Program<F>] which are likely to be correspond to significant inputs during fuzzing.

build_dictionary_from_unconstrained_function 🔒

Collect Field values used in the opcodes of a Brillig function.