noir_greybox_fuzzer::mutation
Module dictionary

This file contains the collection of objects for providing program-specific values to the mutator

Structs

FullDictionary

An object with values from the program used for mutating inputs Contains an int dictionary, where elements are grouped into appropriate widths

IntDictionary

A dictionary for integer values. Separated by width

Functions

add_elements_from_input_map_to_vector_without_abi

Parse input map and add elements to the dictionary We use this when the ABI for a harness has changed and we can no longer directly decode previous testcases, but want to use interesting values from previous testcases in the fuzzing campaign

add_elements_from_input_value_to_vector 🔒

Parse input value and add elements to the dictionary We use this when the ABI for a harness has changed and we can no longer directly decode previous testcases, but want to use interesting values from previous testcases in the fuzzing campaign

Type Aliases

FieldDictionary