noir_greybox_fuzzer
Struct FuzzedExecutor

pub struct FuzzedExecutor<E, F> {

    pub(crate) acir_program: ProgramArtifact,
    pub(crate) brillig_program: ProgramArtifact,
    pub(crate) acir_executor: E,
    pub(crate) brillig_executor: F,
    pub(crate) location_to_feature_map: HashMap<(usize, usize), usize>,
    pub(crate) brillig_coverage_ranges: Vec<BrilligCoverageItemRange>,
    pub(crate) mutator: InputMutator,
    pub(crate) package_name: String,
    pub(crate) function_name: String,
    pub(crate) num_threads: usize,
    pub(crate) show_progress: bool,
    pub(crate) failure_configuration: FuzzedExecutorFailureConfiguration,
    pub(crate) corpus_dir: PathBuf,
    pub(crate) minimize_corpus: bool,
    pub(crate) minimized_corpus_dir: PathBuf,
    pub(crate) metrics: Metrics,
    pub(crate) timeout: u64,
    pub(crate) max_executions: usize,

}

An executor for Noir programs which which provides fuzzing support

After instantiation, calling fuzz will proceed to hammer the program with inputs, until it finds a counterexample.

Fields
acir_program: ProgramArtifact

The program to be fuzzed (acir version)

brillig_program: ProgramArtifact

The program to be fuzzed (brillig version)

acir_executor: E

A function which executes the programs with a given set of inputs

brillig_executor: F

A function which executes the programs with a given set of inputs

location_to_feature_map: HashMap<(usize, usize), usize>

Location to feature map (used in brillig fuzzing)

brillig_coverage_ranges: Vec<BrilligCoverageItemRange>

Brillig coverage ranges (which are branch coverage and which are comparison coverage)

mutator: InputMutator

The object generating mutated version of testcases in the corpus

package_name: String

The name of the package being fuzzed

function_name: String

The name of the function being fuzzed

num_threads: usize

Number of threads to use

show_progress: bool

Whether to output progress to stdout or not.

failure_configuration: FuzzedExecutorFailureConfiguration

Determines what is considered a failure during execution

corpus_dir: PathBuf

Corpus folder

minimize_corpus: bool

If this is set, perform minimization of the corpus

minimized_corpus_dir: PathBuf

Corpus with the minimized

metrics: Metrics

Execution metric

timeout: u64

Maximum time in seconds to spend fuzzing (default: no timeout)

max_executions: usize

Maximum number of executions of ACIR and Brillig (default: no limit)

Implementations

impl<E: Fn(&Program<FieldElement>, WitnessMap<FieldElement>) -> Result<WitnessStack<FieldElement>, ErrorAndWitness> + Sync, F: Fn(&Program<FieldElement>, WitnessMap<FieldElement>, &HashMap<(usize, usize), usize>) -> Result<WitnessAndCoverage, ErrorAndCoverage> + Sync> FuzzedExecutor<E, F>

pub fn new( acir_and_brillig_programs: AcirAndBrilligPrograms, acir_executor: E, brillig_executor: F, package_name: &str, function_name: &str, fuzz_execution_config: FuzzedExecutorExecutionConfiguration, failure_configuration: FuzzedExecutorFailureConfiguration, folder_configuration: FuzzedExecutorFolderConfiguration, ) -> Self

Instantiates a fuzzed executor given an executor

pub(crate) fn detect_new_coverage_from_witness_and_brillig( accumulated_coverage: &AccumulatedFuzzerCoverage, witness: &Option<WitnessStack<FieldElement>>, brillig_coverage: &Option<Vec<u32>>, ) -> bool

Given the witness from ACIR execution and coverage from Brillig execution, check if they have any new coverage ACIR witness is optional, since we can skip ACIR execution or it could have failed, but we collected coverage from brillig We never expect brillig to have no coverage (at least for now)

pub(crate) fn filter_starting_corpus( &self, corpus: &Corpus, starting_corpus_ids: &mut Vec<u64>, ) -> Option<Vec<FieldElement>>

Filter the starting corpus and add elements to the dictionary Removes testcases that can’t be encoded by the new ABI and adds interesting values from them to the dictionary

pub fn fuzz(&mut self) -> FuzzTestResult

Start the fuzzing campaign

pub fn single_fuzz_acir_and_brillig( &self, testcase: &TestCase<'_>, ) -> HarnessExecutionOutcome

Execute acir and brillig programs with the following Testcase

pub(crate) fn handle_failed_case( &self, err: &str, witness: Option<WitnessStack<FieldElement>>, coverage: Option<Vec<u32>>, acir_elapsed: Duration, brillig_elapsed: Duration, testcase: &TestCase<'_>, ) -> HarnessExecutionOutcome

Handle a failed execution case The handling depends on the failure configuration

pub fn single_fuzz_acir( &self, testcase: &TestCase<'_>, ) -> HarnessExecutionOutcome

Granular and single-step function that runs only one fuzz and returns either a CaseOutcome or a CounterExampleOutcome

pub fn single_fuzz_brillig( &self, testcase: &TestCase<'_>, ) -> HarnessExecutionOutcome

Granular and single-step function that runs only one fuzz and returns either a CaseOutcome or a CounterExampleOutcome