noir_greybox_fuzzer
Struct Metrics

pub(crate) struct Metrics {

    pub(crate) total_acir_execution_time: u128,
    pub(crate) total_brillig_execution_time: u128,
    pub(crate) total_mutation_time: u128,
    pub(crate) processed_testcase_count: usize,
    pub(crate) removed_testcase_count: usize,
    pub(crate) removed_testcase_last_round: bool,
    pub(crate) active_corpus_size: usize,
    pub(crate) last_round_size: usize,
    pub(crate) last_round_execution_time: u128,
    pub(crate) last_round_update_time: u128,
    pub(crate) num_threads: usize,
    pub(crate) acir_brillig_discoveries: usize,
    pub(crate) found_new_with_acir_brillig: bool,
    pub(crate) brillig_discoveries: usize,
    pub(crate) found_new_with_brillig: bool,
    pub(crate) pulse_interval_millis: u64,

}

The metrics of the fuzzing process being output to the user

Fields
total_acir_execution_time: u128

Total time spent executing ACIR programs in microseconds

total_brillig_execution_time: u128

Total time spent executing Brillig programs in microseconds

total_mutation_time: u128

Total time spent mutating testcases

processed_testcase_count: usize

The number of unique testcases run

removed_testcase_count: usize

Number of testcases removed from the corpus

removed_testcase_last_round: bool

Purged a testcase last round

active_corpus_size: usize

The size of the corpus being used in mutation schedule

last_round_size: usize

Last round size

last_round_execution_time: u128

Last round execution time

last_round_update_time: u128

Last round accumulated coverage update time

num_threads: usize

Number of threads involved in fuzzing

acir_brillig_discoveries: usize

Number of testcases discovered with ACIR/Brillig tandem

found_new_with_acir_brillig: bool

Discovered something with ACIR/Brillig tandem last round

brillig_discoveries: usize

Number of testcases discovered with Brillig

found_new_with_brillig: bool

Discovered something with Brillig last round

pulse_interval_millis: u64

Pulse interval in milliseconds

Implementations

impl Metrics

pub(crate) fn increase_total_acir_duration_micros(&mut self, update: &u128)

pub(crate) fn increase_total_brillig_duration_micros(&mut self, update: &u128)

pub(crate) fn increase_total_mutation_time(&mut self, update: &u128)

pub(crate) fn is_brillig_dominating(&self) -> bool

Tells if more time has been spent in brillig execution than in ACIR execution

pub(crate) fn increase_processed_testcase_count(&mut self, update: &usize)

pub(crate) fn increment_removed_testcase_count(&mut self)

pub(crate) fn set_active_corpus_size(&mut self, new_size: usize)

pub(crate) fn set_last_round_size(&mut self, new_size: usize)

pub(crate) fn set_last_round_execution_time(&mut self, new_time: u128)

pub(crate) fn set_last_round_update_time(&mut self, new_time: u128)

pub(crate) fn set_num_threads(&mut self, count: usize)

pub(crate) fn increment_acir_brillig_discoveries(&mut self)

pub(crate) fn increment_brillig_discoveries(&mut self)

pub(crate) fn refresh_round(&mut self)