noir_greybox_fuzzer

Module types

Structs

CounterExampleOutcome

Returned by a single fuzz when a counterexample has been discovered

DiscrepancyOutcome

Returned by a single fuzz when there is a discrepancy between brillig and acir execution

ForeignCallErrorInFuzzing

Foreign Call issues

ProgramFailureResult

Returned to the fuzz op in case the fuzzer found a failure case

SuccessfulCaseOutcome

Returned by a single fuzz in the case of a successful run

Enums

FuzzTestResult

The outcome of a fuzz test

HarnessExecutionOutcome

Outcome of a single execution (Brillig or Acir and Brillig)

Type Aliases

CounterExample 🔒