noir_greybox_fuzzer
Module corpus

Structs

Corpus

Main corpus manager that tracks all discovered testcases and orchestrates their selection for fuzzing

CorpusFileManager

Manages reading and writing corpus files to the filesystem, including parsing JSON inputs

Sequence 🔒

Tracks information about a sequence of executions for a particular testcase, including how many executions remain and which testcase is being used

TestCase

Represents a single test case in the corpus, containing both the input values and a unique identifier

TestCaseOrchestrator

Manages the selection of testcases for fuzzing, tracking execution counts and scheduling

Constants

CORPUS_FILE_EXTENSION 🔒

DEFAULT_CORPUS_FOLDER

Statics

NEXT_TESTCASE_ID 🔒

Functions

generate_testcase_id 🔒

Generates a unique sequential ID for a testcase by atomically incrementing a global counter

Type Aliases

NextSelection 🔒

Represents the next testcase(s) selected for fuzzing:

TestCaseId