noir_greybox_fuzzer
Module mutation

Re-exports

pub use dictionary::add_elements_from_input_map_to_vector_without_abi;

Modules

array 🔒

configurations 🔒

This file contains configurations for selecting particular behaviors during mutations

dictionary 🔒

This file contains the collection of objects for providing program-specific values to the mutator

field 🔒

This file contains mechanisms for deterministically mutating a given field value Types of mutations applied:

int 🔒

This file contains mechanisms for mutating integer InputValues. If the value is a boolean, it just picks a new random value. Otherwise, it performs one of the following mutations:

string 🔒

Structs

InputMutator

NodeWeight

NodeWeight determines the probability of mutating a particular object It represent the relative weight of this InputValue and its children

Constants

BOOL_WEIGHT 🔒

FIELD_WEIGHT 🔒

MUTATION_LOG_MAX 🔒

MUTATION_LOG_MIN 🔒

STRING_WEIGHT_MULTIPLIER 🔒